Commodity部分实现o(╯□╰)o
实在是写不动了= =精神开始恍惚了o(╯□╰)o

+++ JXCSystemClient/src/main/java/po/CommodityPO.java
@@ -44,7 +44,7 @@ public class CommodityPO implements Serializable{
 		this.warn=0;
 	}
 	
-	public CommodityPO(String name,String type,int in_price,int out_price){
+	public CommodityPO(String name,String type,double in_price,double out_price){
 		this.name = name;
 		this.type = type;
 		this.in_price = in_price;
